Refactor DTOs, add shipping/payment & promo
Reorganize and extend order/cart domain: move cart DTOs into dtos.cart; introduce paymentTransactions, promocode and order DTO packages; replace simple numeric price types with BigDecimal/long where appropriate. Rename ShipOptions -> ShippingOptions (entities, DTOs, repo, controller) and adjust related types. Add ItemTracking and ShippingAddress endpoints, services, entities, repositories and mappers; extend ItemTracking and Order entities with ETA/ATA and validation. Revamp PaymentTransactions flow: new request/response DTOs, mapper method, integrate cart & promo services to compute totals (BigDecimal) and persist detailed payment fields. Update PromoCodeService to return structured discount results. Update CartService to calculate order prices using BigDecimal and return new OrderPricesDTO. Also apply related validation and field renames, and propagate DTO/package name changes across mappers/controllers. Misc: product price type changed to long and several user-service address/card classes renamed.

+++ productservice/src/main/java/com/khapara/productservice/dtos/ProductDTO.java
@@ -14,7 +14,7 @@ public class ProductDTO {
     private String groupId;
     private String productBrand;
     private String productTagline;
-    private Double productPrice;
+    private long productPrice;
     private Integer productStock;
     private String productColorName;
     private String productColorHex;
