+++ productservice/src/main/java/com/khapara/productservice/entities/Product.java
@@ -43,7 +43,7 @@ public class Product {
 
     @NotNull(message = "Price is required")
     @DecimalMin(value = "0.0", inclusive = false, message = "Price must be greater than 0")
-    private Double productPrice;
+    private long productPrice;
 
     @NotNull(message = "Stock is required")
     @Min(value = 0, message = "Stock cannot be negative")
